requirements-server.txt
gTTS
SpeechRecognition
numpy
pygame
pydub
ffmpeg-python
python-dotenv
openai
pyaudio
vosk